+++ mycodo/utils/update_dependencies.py
@@ -7,28 +7,18 @@ sys.path.append(
     os.path.abspath(os.path.join(
         os.path.dirname(__file__), os.path.pardir) + '/..'))
 
-from mycodo.config import CAMERA_INFO
-from mycodo.config import DEPENDENCIES_GENERAL
-from mycodo.config import FUNCTION_INFO
-from mycodo.config import INSTALL_DIRECTORY
-from mycodo.config import DEPENDENCY_LOG_FILE
-from mycodo.config import METHOD_INFO
-from mycodo.databases.models import Actions
-from mycodo.databases.models import Widget
-from mycodo.databases.models import Camera
-from mycodo.databases.models import Trigger
-from mycodo.databases.models import CustomController
-from mycodo.databases.models import EnergyUsage
-from mycodo.databases.models import Function
-from mycodo.databases.models import Input
-from mycodo.databases.models import Method
-from mycodo.databases.models import Output
+from mycodo.config import (CAMERA_INFO, DEPENDENCIES_GENERAL,
+                           DEPENDENCY_LOG_FILE, FUNCTION_INFO,
+                           INSTALL_DIRECTORY, METHOD_INFO)
+from mycodo.databases.models import (Actions, Camera, Conditional,
+                                     CustomController, EnergyUsage, Function,
+                                     Input, Method, Output, Trigger, Widget)
 from mycodo.mycodo_flask.utils.utils_general import return_dependencies
-from mycodo.utils.functions import parse_function_information
 from mycodo.utils.actions import parse_action_information
 from mycodo.utils.database import db_retrieve_table_daemon
-from mycodo.utils.outputs import parse_output_information
+from mycodo.utils.functions import parse_function_information
 from mycodo.utils.inputs import parse_input_information
+from mycodo.utils.outputs import parse_output_information
 from mycodo.utils.system_pi import cmd_output
 from mycodo.utils.widgets import parse_widget_information
 
@@ -94,6 +84,10 @@ if __name__ == "__main__":
     for each_dev in function:
         if each_dev.function_type not in devices:
             devices.append(each_dev.function_type)
+    
+    conditional = db_retrieve_table_daemon(Conditional, entry='first')
+    if conditional and 'conditional_conditional' not in devices:
+        devices.append('conditional_conditional')
 
     trigger = db_retrieve_table_daemon(Trigger)
     for each_dev in trigger: